Remove WAR Gradle plugin and add project-report plugin
We remove the WAR Gradle plugin because we will run the application container-less from a jar file. This makes running the app much more easy, since we don't need a Tomcat installation. The runnable jar will use an embedded Tomcat container and contain everything it needs to run, so no need to build it as a war file and deploy it anywhere.