Fixed problem with specifying process arguments on linux
Somehow it didn't like the double quotes on linux. To be seen what we can do if the paths contains whitespaces at some point. On Windows the path should probably be wrapped in double quotes. On linux it might require escaping " " with "\ ", but not sure. In any case, linux didn't work with a double quotes classpath argument.