Modify match execution error reporting in order to not expose to much sensitive information (file paths) to the users
When the match execution fails, the users can see an error message in their failed matches. This error message should not contain sensitive information like file paths, but be more general. The information can be retrieved by the admin in the server logs, if necessary.

When matches fail, it's usually due to failed bot instantiation or timeouts. These messages will be communicated to the user and don't reveal sensitive information.