Refactor the user creation mechanism to use a separate DTO and introduce a competition registration password for new user registration
The User is a quite sensitive object, so we should not use it in requests and views. It's good security practice to use a separate DTO class for that. Also, it allows to split out the validation annotations and give use more flexibility with the sign up form by allows custom fields that are not stored in the database (registration password).