Rearrange the packages of the core module in order to enforce package-private visibility of core classes
Merging all of the core classes (except test bots) under the common package "cern.ais.gridwars" is not so nice in terms of clarity, but it's currently the only way how we can minimize the visibility of the core classes to this package. We can use package-private classes and method in order to shield the core classes as much as possible from the bot code. If the bot code does not live in the same "cern.ais.gridwars" package, they will not see most of the classes. And then, if they see them, they won't be able to get the objects at runtime because all the classes are already privatised and pretty locked down. But the more encapsulation, the better.